+++ react-app/server/invoicing.js
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Page, Text, View, Document, StyleSheet, ReactPDF } from '@react-pdf/renderer';
 
+//TODO css styling for invoice pdf
 const styles = StyleSheet.create({
     page: {},
     name: {},
@@ -11,17 +12,19 @@ const styles = StyleSheet.create({
     paidCash: {},
 });
 
+//takes values from order form and renders invoice pdf to a file that can be sent by email
 function genInvoice(order){
     const paidCash = false;
-    if(order.payType=='cash') paidCash=true;
+    if(order.payment=='invoice') paidCash=true;
     
     const nameDisp = 'Name: '+order.firstName+' '+order.lastName;
     const priceDisp = 'Total: $'+(order.numBags*bagPrice);
-    const payTypeDisp = 'Payment Method: '+order.payType;
-    const pickupLocDisp = 'Pickup Address: '+order.pickupLoc;
+    const payTypeDisp = 'Payment Method: '+order.payment;
+    const pickupLocDisp = 'Pickup Address: '+order.pickUp;
     const paidCashDisp = paidCash ? 'CASH PAYMENT DUE ON PICKUP' : '';
 
-    const invoice = () => (
+    //using React-PDF, pdf is treated as react component and can be styled with css
+    const invoiceDoc = () => (
         <Document>
             <Page size='A4' style={styles.page}>
                 <View style={styles.name}>
@@ -43,5 +46,6 @@ function genInvoice(order){
         </Document>
     );
 
-    ReactPDF.render(<invoice />, `${__dirname}/example.pdf`);
+    //TODO come up with unique naming so pdfs can be attached to email and subsequently deleted but so concurrent orders aren't confused
+    ReactPDF.render(<invoiceDoc />, `${__dirname}/example.pdf`);
 }
